demochatservice/requirements.txt
flask==3.0.0
Flask-SQLAlchemy
SQLAlchemy-serializer
pyjwt
werkzeug